+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/IUncrafterContainer.java
@@ -5,6 +5,7 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.RecipeManager;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 public interface IUncrafterContainer {
@@ -13,10 +14,15 @@ public interface IUncrafterContainer {
 
 	OutputHandler getOutputHandler();
 
+	@Nullable
+	EnchantmentHandler getEnchantmentHandler();
+
 	RecipeManager getRecipeManager();
 
 	Tuple<Item, List<ItemStack>> getCache();
 
 	boolean isInputLocked();
+
+	boolean isAdvanced();
 
 }