+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 ### Added
+- Advanced Uncrafter. Allow three ingredients to be retrieved. Allow three enchantments to be retrieved (as an enchanted book).
 - When extracting an ingredient from the output slots, the input slot is locked until you have taken all possible ingredients.
 A little cross appears on the input slot when it is locked.
 
 ### Changed
-- Changed the texture again (it's better now).
+- Changed the texture for the uncrafter again (it's better now).
 
 ### Fixed
 - Uncrafter having end stone as bottom texture instead of crying obsidian